+++ internal/types.go
@@ -1,24 +1,10 @@
 package internal
 
-import (
-	"go/token"
-	"os"
-	"strings"
-)
+import "go/token"
 
 // SourceCode stores the content of a source code file.
 type SourceCode struct {
 	Lines []string
-}
-
-// ReadSourceFile reads the content of a file and returns it as a `SourceCode` struct.
-func ReadSourceCode(filename string) (*SourceCode, error) {
-	content, err := os.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-	lines := strings.Split(string(content), "\n")
-	return &SourceCode{Lines: lines}, nil
 }
 
 // Issue represents a lint issue found in the code base.
